rag-service/requirements.txt
pydantic>=2.0.0
loguru>=0.7.0
httpx
fastapi>=0.95.0
uvicorn>=0.21.0
weaviate-client==3.26.7
langchain>=0.0.175
langchain-huggingface>=0.0.7
langchain-community>=0.1.0
sentence-transformers

openai==1.57.4
